+++ src/platform.ts
@@ -129,8 +129,19 @@ export class ShellyPlatform extends MatterbridgeDynamicPlatform {
     // handle Shelly discovered event
     this.shelly.on('discovered', async (discoveredDevice: DiscoveredDevice) => {
       if (this.discoveredDevices.has(discoveredDevice.id)) {
-        this.log.info(`Shelly device ${hk}${discoveredDevice.id}${nf} host ${zb}${discoveredDevice.host}${nf} already discovered`);
-        return;
+        const stored = this.storedDevices.get(discoveredDevice.id);
+        if (stored?.host !== discoveredDevice.host) {
+          this.log.warn(`Shelly device ${hk}${discoveredDevice.id}${wr} host ${zb}${discoveredDevice.host}${wr} is already discovered with a different host.`);
+          this.log.warn(`Set new address for shelly device ${hk}${discoveredDevice.id}${wr} from ${zb}${stored?.host}${wr} to ${zb}${discoveredDevice.host}${wr}`);
+          this.log.warn(`Please restart for the change to take effect.`);
+          this.discoveredDevices.set(discoveredDevice.id, discoveredDevice);
+          this.storedDevices.set(discoveredDevice.id, discoveredDevice);
+          await this.saveStoredDevices();
+          return;
+        } else {
+          this.log.info(`Shelly device ${hk}${discoveredDevice.id}${nf} host ${zb}${discoveredDevice.host}${nf} already discovered`);
+          return;
+        }
       }
       this.discoveredDevices.set(discoveredDevice.id, discoveredDevice);
       this.storedDevices.set(discoveredDevice.id, discoveredDevice);